hide placeholder projects
- doesn't look professional
- will reinstate them once I'm done working on real projects
