+++ components/Contact/index.tsx
@@ -1,44 +1,17 @@
-import { useState } from "react";
-import { useForm } from "react-hook-form";
 import { FiCheckCircle, FiInfo } from "react-icons/fi";
 import Container from "../Container";
 import InputField from "./InputField";
+import useContact from "./useContact";
 
 function Contact() {
   const {
     register,
-    handleSubmit,
-    formState: { errors, isSubmitting, isSubmitSuccessful },
-  } = useForm<FormInput>();
-
-  const [error, setError] = useState("");
-
-  const onSubmit = async (data: FormInput) => {
-    const response = await fetch("/api/sendgrid", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(data),
-    });
-
-    const { error } = await response.json();
-    setError(error);
-  };
-
-  const SubmitFailed = (
-    <>
-      <FiInfo />
-      {error}
-    </>
-  );
-
-  const SubmitSucceeded = (
-    <>
-      <FiCheckCircle />
-      {"Email sent! I'll get back to you soon."}
-    </>
-  );
+    submit,
+    errors,
+    serverError,
+    isSubmitting,
+    isSubmitSuccessful,
+  } = useContact();
 
   return (
     <Container
@@ -50,16 +23,7 @@ function Contact() {
       </h2>
       <p className="mt-2">{"Let's get in touch"}</p>
 
-      <form
-        className="mt-16 space-y-8"
-        onSubmit={
-          isSubmitting
-            ? (e) => {
-                e.preventDefault();
-              }
-            : handleSubmit(onSubmit)
-        }
-      >
+      <form className="mt-16 space-y-8" onSubmit={submit}>
         <InputField
           label="name"
           placeholder="Leonardo"
@@ -83,25 +47,35 @@ function Contact() {
           error={errors.message}
         />
 
-        {isSubmitSuccessful ? (
+        <button
+          type="submit"
+          aria-disabled={isSubmitting}
+          className={`px-12 py-2 font-semibold transition-opacity bg-black rounded-md text-slate-50 hover:bg-opacity-80 ${
+            isSubmitting ? "cursor-not-allowed" : ""
+          }`}
+        >
+          {isSubmitting ? "Sending..." : "Send"}
+        </button>
+
+        {isSubmitSuccessful && (
           <div
             role="alert"
             className={`flex items-center gap-2 px-4 py-2 w-[var(--input-width)] rounded-md md:text-base text-white ${
-              error ? "bg-red-600" : "bg-green-600"
+              serverError ? "bg-red-600" : "bg-green-600"
             }`}
           >
-            {error ? SubmitFailed : SubmitSucceeded}
+            {serverError ? (
+              <>
+                <FiInfo />
+                {serverError}
+              </>
+            ) : (
+              <>
+                <FiCheckCircle />
+                {"Email sent! I'll get back to you soon."}
+              </>
+            )}
           </div>
-        ) : (
-          <button
-            type="submit"
-            aria-disabled={isSubmitting}
-            className={`px-12 py-2 font-semibold transition-opacity bg-black rounded-md text-slate-50 hover:bg-opacity-80 ${
-              isSubmitting ? "cursor-not-allowed" : ""
-            }`}
-          >
-            {isSubmitting ? "Sending..." : "Send"}
-          </button>
         )}
       </form>
     </Container>
